convert clonedSubmissionData to unsavedSubmissionData local state
