+++ app/frontend/components/shared/permit-applications/requirement-form.tsx
@@ -18,7 +18,7 @@ import { observer } from "mobx-react-lite"
 
 import { format } from "date-fns"
 import * as R from "ramda"
-import React, { useEffect, useMemo, useRef, useState } from "react"
+import React, { useEffect, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { useLocation, useNavigate } from "react-router-dom"
 import { useMountStatus } from "../../../hooks/use-mount-status"
@@ -43,14 +43,7 @@ interface IRequirementFormProps {
 }
 
 export const RequirementForm = observer(
-  ({
-    permitApplication,
-    onCompletedBlocksChange,
-    formRef,
-    triggerSave,
-    showHelpButton = true,
-    isEditing = false,
-  }: IRequirementFormProps) => {
+  ({ permitApplication, onCompletedBlocksChange, formRef, triggerSave, isEditing = false }: IRequirementFormProps) => {
     const {
       jurisdiction,
       submissionData,
@@ -77,7 +70,12 @@ export const RequirementForm = observer(
     const [firstComponentKey, setFirstComponentKey] = useState(null)
     const [isCollapsedAll, setIsCollapsedAllState] = useState(false)
 
-    const clonedSubmissionData = useMemo(() => R.clone(submissionData), [submissionData])
+    const [unsavedSubmissionData, setUnsavedSubmissionData] = useState(R.clone(submissionData))
+
+    const handleSetUnsavedSubmissionData = (data) => {
+      permitApplication.setIsDirty(true)
+      setUnsavedSubmissionData(data)
+    }
 
     const { isOpen: isContactsOpen, onOpen: onContactsOpen, onClose: onContactsClose } = useDisclosure()
 
@@ -357,7 +355,7 @@ export const RequirementForm = observer(
             formReady={formReady}
             /* Needs cloned submissionData otherwise it's not possible to use data grid as mst props
             can't be mutated*/
-            submission={clonedSubmissionData}
+            submission={unsavedSubmissionData}
             onSubmit={onFormSubmit}
             options={permitAppOptions}
             onBlur={onBlur}
@@ -410,7 +408,8 @@ export const RequirementForm = observer(
             onClose={onContactsClose}
             autofillContactKey={autofillContactKey}
             permitApplication={permitApplication}
-            submissionState={submissionData}
+            submissionState={unsavedSubmissionData}
+            setSubmissionState={handleSetUnsavedSubmissionData}
           />
         )}
       </>
